finish example + test + add missing files
