+++ workspace-android/.gitignore
@@ -1,11 +1,14 @@
 *.iml
 .gradle
 /local.properties
-.idea/
+/.idea/caches
+/.idea/libraries
+/.idea/modules.xml
+/.idea/workspace.xml
+/.idea/navEditor.xml
+/.idea/assetWizardSettings.xml
 .DS_Store
 /build
 /captures
 .externalNativeBuild
 .cxx
-
-scripts/build/
