+++ workspace-android/.idea/.name
@@ -1,0 +1,1 @@
+My Application
