+++ workspace-android/app/src/androidTest/java/io/techhq/myapplication/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package io.techhq.myexample
+package io.techhq.myapplication
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("io.techhq.myexample", appContext.packageName)
+        assertEquals("io.techhq.myapplication", appContext.packageName)
     }
 }
