+++ workspace-android/app/src/main/java/io/techhq/myapplication/MainActivity.kt
@@ -1,0 +1,15 @@
+package io.techhq.myapplication
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import io.techhq.myexample.MySimpleInteraction
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        val msi = MySimpleInteraction()
+        msi.connectAndCall()
+    }
+}
