+++ workspace-android/app/src/main/java/io/techhq/myapplication/MySimpleInteraction.kt
@@ -1,8 +1,6 @@
 package io.techhq.myexample
 
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import io.techhq.myexample.contracts.SimpleStorage
+import io.techhq.myapplication.contracts.SimpleStorage
 import org.web3j.crypto.Credentials
 import org.web3j.crypto.WalletUtils
 import org.web3j.protocol.Web3j
@@ -10,13 +8,8 @@ import org.web3j.protocol.http.HttpService
 import org.web3j.tx.gas.DefaultGasProvider
 import java.io.File
 
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
+class MySimpleInteraction {
+    fun connectAndCall(): Long {
         val web3 = Web3j.build(HttpService("http://localhost:8545"))
         val web3ClientVersion = web3.web3ClientVersion().send()
         val clientVersion = web3ClientVersion.web3ClientVersion
@@ -30,8 +23,10 @@ class MainActivity : AppCompatActivity() {
                     File.separator,
                     File.separator
                 ))
+
         val contract: SimpleStorage = SimpleStorage.load(
-                "0x<address>|<ensName>", web3, credentials, DefaultGasProvider())
-        println(contract.get().send())
+            "0xf2Dee5975A808f16f93bf4Fd55aB5481a8B20497", web3, credentials, DefaultGasProvider()
+        )
+        return contract.get().send().toLong()
     }
 }
