+++ workspace-android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">My Example</string>
+    <string name="app_name">My Application</string>
 </resources>
