+++ workspace-android/gradle/wrapper/gradle-wrapper.properties
@@ -1,4 +1,4 @@
-#Tue Jan 07 17:32:28 WET 2020
+#Fri Jan 10 19:39:25 WET 2020
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 zipStoreBase=GRADLE_USER_HOME
