+++ workspace-android/settings.gradle
@@ -1,2 +1,2 @@
 include ':app'
-rootProject.name='My Example'
+rootProject.name='My Application'
